Rename some more files
