+++ flopy4/mf6/modflow/slnims.py
