+++ test/test_interface.py
@@ -7,9 +7,9 @@ from flopy.discretization.modeltime import ModelTime
 
 from flopy4.mf6.adapters import Flopy3Model, Flopy3Package
 from flopy4.mf6.modflow.gwf import Chd, Dis, Gwf, Ic, Npf, Oc
-from flopy4.mf6.modflow.ims import Ims
-from flopy4.mf6.modflow.simulation import Simulation
-from flopy4.mf6.modflow.tdis import Tdis
+from flopy4.mf6.modflow.sim import Simulation
+from flopy4.mf6.modflow.simtdis import Tdis
+from flopy4.mf6.modflow.slnims import Ims
 
 
 def quickstart_model():
